+++ src/xqt-kotlinx-lsp/commonTest/xqt/kotlinx/lsp/test/types/TextDocumentPositionParams.kt
@@ -1,10 +1,10 @@
-// Copyright (C) 2022 Reece H. Dunn. SPDX-License-Identifier: Apache-2.0
+// Copyright (C) 2022-2023 Reece H. Dunn. SPDX-License-Identifier: Apache-2.0
 package xqt.kotlinx.lsp.test.types
 
 import kotlinx.serialization.json.JsonArray
 import kotlinx.serialization.json.JsonNull
 import kotlinx.serialization.json.JsonPrimitive
-import xqt.kotlinx.lsp.types.TextDocumentPosition
+import xqt.kotlinx.lsp.types.TextDocumentPositionParams
 import xqt.kotlinx.rpc.json.serialization.UnsupportedKindTypeException
 import xqt.kotlinx.rpc.json.serialization.jsonObjectOf
 import xqt.kotlinx.test.DisplayName
@@ -12,8 +12,8 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFails
 
-@DisplayName("The TextDocumentPosition type")
-class TheTextDocumentPositionType {
+@DisplayName("The TextDocumentPositionParams type")
+class TheTextDocumentPositionParamsType {
     @Test
     @DisplayName("supports the non-optional properties")
     fun supports_the_non_optional_properties() {
@@ -25,38 +25,38 @@ class TheTextDocumentPositionType {
             )
         )
 
-        val tdpp = TextDocumentPosition.deserialize(json)
+        val tdpp = TextDocumentPositionParams.deserialize(json)
         assertEquals("file:///home/lorem/ipsum.py", tdpp.uri)
         assertEquals(2u, tdpp.position.line)
         assertEquals(6u, tdpp.position.character)
 
-        assertEquals(json, TextDocumentPosition.serializeToJson(tdpp))
+        assertEquals(json, TextDocumentPositionParams.serializeToJson(tdpp))
     }
 
     @Test
     @DisplayName("throws an error if the kind type is not supported")
     fun throws_an_error_if_the_kind_type_is_not_supported() {
-        val e1 = assertFails { TextDocumentPosition.deserialize(JsonArray(arrayListOf())) }
+        val e1 = assertFails { TextDocumentPositionParams.deserialize(JsonArray(arrayListOf())) }
         assertEquals(UnsupportedKindTypeException::class, e1::class)
         assertEquals("Unsupported kind type 'array'", e1.message)
 
-        val e2 = assertFails { TextDocumentPosition.deserialize(JsonNull) }
+        val e2 = assertFails { TextDocumentPositionParams.deserialize(JsonNull) }
         assertEquals(UnsupportedKindTypeException::class, e2::class)
         assertEquals("Unsupported kind type 'null'", e2.message)
 
-        val e3 = assertFails { TextDocumentPosition.deserialize(JsonPrimitive("test")) }
+        val e3 = assertFails { TextDocumentPositionParams.deserialize(JsonPrimitive("test")) }
         assertEquals(UnsupportedKindTypeException::class, e3::class)
         assertEquals("Unsupported kind type 'string'", e3.message)
 
-        val e4 = assertFails { TextDocumentPosition.deserialize(JsonPrimitive(true)) }
+        val e4 = assertFails { TextDocumentPositionParams.deserialize(JsonPrimitive(true)) }
         assertEquals(UnsupportedKindTypeException::class, e4::class)
         assertEquals("Unsupported kind type 'boolean'", e4.message)
 
-        val e5 = assertFails { TextDocumentPosition.deserialize(JsonPrimitive(1)) }
+        val e5 = assertFails { TextDocumentPositionParams.deserialize(JsonPrimitive(1)) }
         assertEquals(UnsupportedKindTypeException::class, e5::class)
         assertEquals("Unsupported kind type 'integer'", e5.message)
 
-        val e6 = assertFails { TextDocumentPosition.deserialize(JsonPrimitive(1.2)) }
+        val e6 = assertFails { TextDocumentPositionParams.deserialize(JsonPrimitive(1.2)) }
         assertEquals(UnsupportedKindTypeException::class, e6::class)
         assertEquals("Unsupported kind type 'decimal'", e6.message)
     }